Fiks varsler når bruker ikke er under oppfølging
