+++ frontend/mulighetsrommet-veileder-flate/src/apps/modia/varsler/BrukerIkkeUnderOppfolgingVarsel.tsx
@@ -6,9 +6,13 @@ interface Props {
 }
 
 export function BrukerIkkeUnderOppfolgingVarsel({ brukerdata }: Props) {
-  return brukerdata.varsler.includes(BrukerVarsel.BRUKER_IKKE_UNDER_OPPFOLGING) ? (
-    <Alert variant="warning" data-testid="varsel_servicesgruppe">
-      Brukeren har ikke fått §14 a-vedtak, og kan derfor ikke meldes på noen tiltak.
-    </Alert>
-  ) : null;
+  if (brukerdata.varsler.includes(BrukerVarsel.BRUKER_IKKE_UNDER_OPPFOLGING)) {
+    return (
+      <Alert variant="warning">
+        Bruker er ikke under oppfølging, og kan derfor ikke meldes på noen tiltak.
+      </Alert>
+    );
+  }
+
+  return null;
 }
